requirements.txt
psycopg2-binary
naoth
numpy
minio
opencv-python
git+https://github.com/HumanSignal/label-studio-sdk@master
pillow
protobuf==3.20
cppyy==3.1.2